Add files via upload
This is the final project report for the MAS416: Modeling and Simulation of Mechatronic Systems course at the University of
Agder. The course is at the Master level and is taken by students in Mechatronics and
Renewable Energy. 
The project includes:
• A strategy for the control of the wind turbine speed including a look-up table with the
desired rotor speed, 𝜔𝑟𝑜𝑡 (𝑟𝑒𝑓), for any wind speed in the interval 0...25 m/s.
• A simulation model of the entire drive train in Matlab/Simscape/Simulink with the control
strategy implemented.